+++ .glitch-assets
@@ -701,3 +701,4 @@
 {"name":"poller_GLITCH.png","date":"2018-11-26T12:14:58.656Z","url":"https://cdn.glitch.com/02863ac1-a499-4a41-ac9c-41792950000f%2Fpoller_GLITCH.png","type":"image/png","size":48543,"imageWidth":2076,"imageHeight":1038,"thumbnail":"https://cdn.glitch.com/02863ac1-a499-4a41-ac9c-41792950000f%2Fthumbnails%2Fpoller_GLITCH.png","thumbnailWidth":330,"thumbnailHeight":165,"dominantColor":"rgb(252,220,44)","uuid":"FS8X0hMcjMCNPMRK"}
 {"name":"conways-webcam_GLITCH.png","date":"2018-11-26T12:14:58.768Z","url":"https://cdn.glitch.com/02863ac1-a499-4a41-ac9c-41792950000f%2Fconways-webcam_GLITCH.png","type":"image/png","size":41372,"imageWidth":2062,"imageHeight":1031,"thumbnail":"https://cdn.glitch.com/02863ac1-a499-4a41-ac9c-41792950000f%2Fthumbnails%2Fconways-webcam_GLITCH.png","thumbnailWidth":330,"thumbnailHeight":165,"dominantColor":"rgb(60,140,84)","uuid":"7HmwuVe7REzkVxtq"}
 {"name":"mouse-photobooth_GLITCH.png","date":"2018-11-26T12:14:58.859Z","url":"https://cdn.glitch.com/02863ac1-a499-4a41-ac9c-41792950000f%2Fmouse-photobooth_GLITCH.png","type":"image/png","size":79412,"imageWidth":2078,"imageHeight":1039,"thumbnail":"https://cdn.glitch.com/02863ac1-a499-4a41-ac9c-41792950000f%2Fthumbnails%2Fmouse-photobooth_GLITCH.png","thumbnailWidth":330,"thumbnailHeight":165,"dominantColor":"rgb(156,84,28)","uuid":"2Wkp58o3xi6VDWsM"}
+{"name":"telescope_404.svg","date":"2018-11-26T18:58:03.849Z","url":"https://cdn.glitch.com/7138972f-76e1-43f4-8ede-84c3cdd4b40a%2Ftelescope_404.svg","type":"image/svg+xml","size":29269,"imageWidth":318,"imageHeight":297,"thumbnail":"https://cdn.glitch.com/7138972f-76e1-43f4-8ede-84c3cdd4b40a%2Ftelescope_404.svg","thumbnailWidth":318,"thumbnailHeight":297,"dominantColor":null,"uuid":"swD7vnauhT5nzYa4"}
