+++ src/presenters/includes/logo.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import moment from 'moment-mini';
+
+class Logo extends React.PureComponent {
+  
+  constructor(props) {
+    super(props);
+    this.state = {
+      hour: (new Date()).getHours(),
+    };
+  }
+  
+  componentDidMount() {
+    this.interval = window.setInterval(() => {
+      this.setState({
+        hour: (new Date()).getHours(),
+      });
+    }, moment.duration(5, 'minutes').asMilliseconds());
+  }
+
+  componentWillUnmount() {
+    window.clearInterval(this.interval);
+  }
+
+  render() {
+    const {hour} = this.state;
+    
+    const LOGO_DAY = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Flogo-day.svg";
+    const LOGO_SUNSET = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Flogo-sunset.svg";
+    const LOGO_NIGHT = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Flogo-night.svg";
+
+    let logo = LOGO_DAY;
+    if ((hour >= 16) && (hour <= 18)) {
+      logo = LOGO_SUNSET;
+    } else if ((hour > 18) || (hour <= 8)) {
+      logo = LOGO_NIGHT;
+    }
+
+    return <img className="logo" src={logo} alt="Glitch" />;
+  }
+}
