+++ src/presenters/pages/404.jsx
@@ -4,9 +4,12 @@ import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import ErrorPage from './error.jsx';
 
+const telescopeImageUrl = 'https://cdn.glitch.com/7138972f-76e1-43f4-8ede-84c3cdd4b40a%2Ftelescope_404.svg?1543258683849';
+
 const NotFoundPage = () => (
   <>
-    <ErrorPage title="Page Not Found" description="Maybe a typo? Or perhaps it's moved?"/>
+    <img src={telescopeImageUrl} alt='' />
+  <div id=''>test</div>
     <Helmet>
       <title>👻 Page not found</title> {/* eslint-disable-line */}
     </Helmet>
