+++ styles/styles.styl
@@ -39,3 +39,4 @@
 @require 'pages/search'
 @require 'pages/error'
 @require 'pages/secret'
+@require 'pages/404'
